refactor: change lib/next-auth as auth
