+++ apps/admin.enes.sh/app/(routes)/api/auth/[...nextauth]/route.ts
@@ -1,4 +1,4 @@
-import { AuthOptions } from '@/packages/next-auth'
+import { AuthOptions } from '@/packages/auth'
 
 import NextAuth from 'next-auth'
